sso_proxy: Add comment marking private key as test data
Despite this key being stored in a `testdata` directory, we've received a number of reports highlighting that a private key is present in the SSO Repo.

PEM format means anything outside of the `--- BEGIN ---` and `-- END ---`` block will be ignored, so it's safe to add a descriptive comment explaining that yes, this really is test data and its intentionally public!